requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python
numpy
scikit-learn
Pillow
requests
joblib 